Refactor: 컴포넌트 네이밍 컨벤션 변경 #20
- PascalCase -> kebab-case
