+++ src/app/careers/page.tsx
@@ -1,5 +1,5 @@
 import { Metadata } from 'next';
-import { Layout } from '@/components/layout/Layout';
+import { CommonLayout } from '@/components/layout/common-layout';
 import Link from 'next/link';
 import { Button } from '@/components/ui/button';
 import { ArrowRight, Building2, Users, Briefcase } from 'lucide-react';
@@ -67,7 +67,7 @@ const positions = [
 
 export default function Careers() {
 	return (
-		<Layout>
+		<CommonLayout>
 			<div className="flex flex-col gap-16 px-4 py-16 sm:px-6 lg:px-8">
 				{/* Hero Section */}
 				<section className="mx-auto flex w-full max-w-7xl flex-col items-center gap-8 text-center">
@@ -160,6 +160,6 @@ export default function Careers() {
 					</Button>
 				</section>
 			</div>
-		</Layout>
+		</CommonLayout>
 	);
 }
