+++ src/app/contact/page.tsx
@@ -1,5 +1,5 @@
 import { Metadata } from 'next';
-import { Layout } from '@/components/layout/Layout';
+import { CommonLayout } from '@/components/layout/common-layout';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { Textarea } from '@/components/ui/textarea';
@@ -68,7 +68,7 @@ const businessTypes = [
 
 export default function Contact() {
 	return (
-		<Layout>
+		<CommonLayout>
 			<div className="flex flex-col gap-16 px-4 py-16 sm:px-6 lg:px-8">
 				{/* Hero Section */}
 				<section className="mx-auto flex w-full max-w-7xl flex-col items-center gap-8 text-center">
@@ -233,6 +233,6 @@ export default function Contact() {
 					</Card>
 				</div>
 			</div>
-		</Layout>
+		</CommonLayout>
 	);
 }
